+++ typescript-client-sdk/src/sdk/models/shared/index.ts
@@ -21,6 +21,7 @@ export * from "./httpbinsimplejsonobject";
 export * from "./limitoffsetconfig";
 export * from "./mapobjvalue";
 export * from "./mapobjvaluecamelcase";
+export * from "./nullableobject";
 export * from "./readonlyobject";
 export * from "./readwriteobject";
 export * from "./refqueryparamobj";