+++ RELEASES.md
@@ -531,3 +531,18 @@ Based on:
 - [go v1.30.4] go-client-sdk
 ### Releases
 - [Go v1.30.4] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v1.30.4 - go-client-sdk
+
+## 2023-09-06 17:00:49
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 
+- Speakeasy CLI 1.78.9 (2.96.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.32.2] typescript-client-sdk
+- [go v1.30.5] go-client-sdk
+- [java v1.31.5] java-client-sdk
+- [php v1.31.5] php-client-sdk
+- [python v1.31.5] python-client-sdk
+- [ruby v1.31.5] ruby-client-sdk
+### Releases
+- [Go v1.30.5] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v1.30.5 - go-client-sdk